added parse Float

+++ src/screens/other/Withdraw.js
@@ -46,7 +46,7 @@ const Withdraw = ({ navigation, route }) => {
       url: ApiUrl.widthdrawalRequest,
       method: 'post',
       body: {
-        amount: parseInt(Money) +  parseInt(processing_fee),
+        amount: parseFloat(Money) +  parseFloat(processing_fee),
         widthdrawl_info: {
           type: bankTab,
           data: bankTab === 'bank' ? [ACName, ACNumber, ACCode] : [PaypalID],
